% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_term.R
\name{create_term}
\alias{create_term}
\title{Create Term for \code{remstats} Effects}
\usage{
create_term(name, prefix, template)
}
\arguments{
\item{name}{A string to process.}

\item{prefix}{The prefix indicating the effect type.}

\item{template}{A template string, where \verb{\%s} represents the extracted variable.}
}
\value{
A formatted string if \code{name} matches \code{prefix}; otherwise, \code{NULL}.
}
\description{
Extracts a variable from a prefixed string and formats it as a \code{remstats} effect term.
}
