% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reh.R
\name{generate_reh}
\alias{generate_reh}
\title{Simulate a Relational Event History (REH)}
\usage{
generate_reh(parameters, covar, M, directed = FALSE)
}
\arguments{
\item{parameters}{A named list of parameters for the model. Each parameter represents
a specific effect or term in the tie-oriented model (e.g., \code{send_z1}, \code{difference_z2},
\code{inertia}). The parameter names must align with those supported by
\code{remreg::generate_formula()}.}

\item{covar}{A data frame of covariates containing at least the following columns:
\itemize{
\item \code{name}: Numeric or character identifiers for actors.
\item \code{z1}: A numeric covariate representing some attribute of the actors.
\item \code{z2}: A numeric covariate representing another attribute of the actors.
\item \code{time}: The starting time for the simulation (default is usually \code{0}).
Each row represents an actor in the network.
}}

\item{M}{An integer specifying the number of events to simulate.}

\item{directed}{A logical value indicating whether the network is directed
(\code{TRUE}) or undirected (\code{FALSE}).}
}
\value{
A data frame representing the simulated event history. The data frame contains
the following columns:
\itemize{
\item \code{time}: The time of each event.
\item \code{actor1}: The identifier of the first actor involved in the event.
\item \code{actor2}: The identifier of the second actor involved in the event.
}
}
\description{
This function simulates a sequence of relational events based on a specified
statistical model derived from a relational event model (REM). It uses covariates,
a list of parameters, and the number of events to generate a relational event history
(REH) with associated statistics.
}
\details{
The \code{generate_reh()} function simulates a series of relational events using the
specified model parameters and covariates. The simulation is based on a relational
event model (REM), a statistical framework for modeling and analyzing dynamic interactions
between actors over time.

It performs the following steps:
\enumerate{
\item \strong{Formula Generation}: The function uses \code{remreg::generate_formula()} to
create a formula for the \code{tie_effects} argument of \code{remstats::remstats()}.
\item \strong{Initialization}: A dummy first event is created to initialize the simulation.
\item \strong{Simulation Loop}: For each event:
\itemize{
\item The model parameters are applied to compute the event rates.
\item A new event is sampled based on the computed rates.
\item The event is added to the event history, and the model is updated.
}
\item \strong{Endogeneity Updates}: The function computes endogeneity statistics for
subsequent events using \code{remstats::remstats()}.
}

\strong{Standardization}: All effects that allow scaling are standardized by default
(\code{scaling = 'std'}), ensuring consistent computation of statistics.
}
\note{
This function is inspired by the work presented in the article:

Lakdawala, R., Mulder, J., & Leenders, R. (2024). \emph{Simulating Relational Event Histories--Why and How}.
arXiv preprint arXiv:2403.19329.
}
\examples{
# Define parameters
parameters <- list(
  baseline = -5,
  send_z1 = 0.5,
  difference_z2 = 0.2,
  inertia = 0.3
)

# Create covariate data
covar <- data.frame(
  name = 1:50, time = 0,
  z1 = rnorm(n = 50, mean = 0, sd = 1),
  z2 = rnorm(n = 50, mean = 0, sd = 1)
)

# Simulate 10 events
events <- generate_reh(parameters, covar, M = 10, directed = TRUE)
print(events$edgelist)

}
\references{
Lakdawala, R., Mulder, J., & Leenders, R. (2024). \emph{Simulating Relational Event Histories--Why and How}.
arXiv preprint arXiv:2403.19329.
}
\seealso{
\code{\link[remstats:remstats]{remstats::remstats()}} for calculating statistics for tie-oriented models.
\code{\link[=generate_formula]{generate_formula()}} for generating the formula used in this function.
}
