% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.R
\name{hdi}
\alias{hdi}
\title{Calculate the Highest Density Interval (HDI)}
\usage{
hdi(sample, prob = 0.95)
}
\arguments{
\item{sample}{A numeric vector representing the sample data.}

\item{prob}{A numeric value between 0 and 1 specifying the desired probability mass. Defaults to 0.95.}
}
\value{
A numeric vector of length 2 indicating the lower and upper bounds of the HDI.
}
\description{
Computes the smallest interval containing the specified probability mass for a given sample.
}
