% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remreg-package.R
\docType{package}
\name{remreg-package}
\alias{remreg}
\alias{remreg-package}
\title{remreg: Generate Relational Event History Data And Evaluate Variable Selection And Predictive Performance}
\description{
This package provides functions to generate relational event history data and evaluate variable selection with approximate Bayesian regularization.
}
\keyword{internal}
